backend/pyproject.toml
[project]
name = "guidetarbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.0",
    "python-jose[cryptography]>=3.5.0",
    "librosa>=0.11.0",
    "madmom",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "soundfile>=0.13.1",
    "uvicorn>=0.35.0",
    "yt-dlp>=2026.3.17",
]

[tool.uv.sources]
madmom = { git = "https://github.com/CPJKU/madmom" }
